fix x64 build

+++ src/oddlib/anim.cpp
@@ -414,7 +414,7 @@ namespace Oddlib
     {
         // Collect all animation sets
         auto hdr = std::make_unique<AnimationHeader>();
-        hdr->mOffset = mStream.Pos();
+        hdr->mOffset = static_cast<u32>(mStream.Pos());
         mStream.Read(hdr->mFps);
         mStream.Read(hdr->mNumFrames);
         mStream.Read(hdr->mLoopStartFrame);
